+++ cli/utils/derivateSS58AccountFromPK.go
@@ -21,18 +21,18 @@ var (
 
 var (
 	privateKey string
-	networkID  uint8
+	networkID  uint16
 )
 
 func init() {
 	derivateSS58AccountFromPKCMD.PersistentFlags().StringVar(&privateKey, "privateKey", "", "hex encoded private key")
 	_ = derivateSS58AccountFromPKCMD.MarkFlagRequired("privateKey")
-	derivateSS58AccountFromPKCMD.PersistentFlags().Uint8Var(&networkID, "networkID", 0, "network id for a checksum. Registry https://github.com/paritytech/ss58-registry/blob/main/ss58-registry.json")
+	derivateSS58AccountFromPKCMD.PersistentFlags().Uint16Var(&networkID, "networkID", 0, "network id for a checksum. Registry https://github.com/paritytech/ss58-registry/blob/main/ss58-registry.json")
 	_ = derivateSS58AccountFromPKCMD.MarkFlagRequired("networkID")
 }
 
 func derivateSS58(cmd *cobra.Command, args []string) error {
-	account, err := signature.KeyringPairFromSecret(privateKey, uint16(networkID))
+	account, err := signature.KeyringPairFromSecret(privateKey, networkID)
 	if err != nil {
 		return err
 	}
